Moved capabilities tests to corresponding packages
